# docs/openapi/smartranking-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/clubs": {
      "post": {
        "operationId": "Clubs_createClub",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClubDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Clubs"
        ]
      },
      "get": {
        "operationId": "Clubs_getAllClubs",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (max 100)",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Case-insensitive search on club name or slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Clubs"
        ]
      }
    },
    "/api/v1/clubs/public": {
      "get": {
        "operationId": "Clubs_getPublicClubs",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (max 100)",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Clubs"
        ]
      }
    },
    "/api/v1/clubs/{_id}": {
      "get": {
        "operationId": "Clubs_getClubById",
        "parameters": [
          {
            "name": "_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Clubs"
        ]
      },
      "put": {
        "operationId": "Clubs_updateClub",
        "parameters": [
          {
            "name": "_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClubDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Clubs"
        ]
      },
      "delete": {
        "operationId": "Clubs_deleteClub",
        "parameters": [
          {
            "name": "_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Clubs"
        ]
      }
    },
...
    "/api/v1/categories": {
      "post": {
        "operationId": "Categories_createCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Categories"
        ]
      },
      "get": {
        "operationId": "Categories_getCategory",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (max 100)",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Case-insensitive filter on category label",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clubId",
            "required": false,
            "in": "query",
            "description": "Restrict categories to a single club",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Categories"
        ]
      }
    },
    "/api/v1/categories/my": {
      "get": {
        "operationId": "Categories_getMyCategories",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (max 100)",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Categories"
        ]
      }
    },
    "/api/v1/categories/{category}": {
      "get": {
        "operationId": "Categories_getCategorieById",
        "parameters": [
          {
            "name": "category",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Categories"
        ]
      },
      "put": {
        "operationId": "Categories_updateCategory",
        "parameters": [
          {
            "name": "category",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Categories"
        ]
      }
    },
    "/api/v1/categories/{category}/players/{playerId}": {
      "post": {
        "operationId": "Categories_assignPlayerCategory",
        "parameters": [
          {
            "name": "category",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "playerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Categories"
        ]
      }
    },
...
  },
  "info": {
    "title": "SmartRanking API",
    "description": "Multi-tenant tennis ranking API with Better Auth, RBAC, and MongoDB.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session-token",
        "description": "Better Auth session token set via /api/auth"
      },
      "Tenant": {
        "type": "apiKey",
        "in": "header",
        "name": "x-tenant-id",
        "description": "Tenant/club identifier. Required for all protected multi-tenant endpoints."
      }
    },
    "schemas": {
      "CreateClubDto": {
        "type": "object",
        "properties": {}
      },
      "Object": {
        "type": "object",
        "properties": {}
      },
      "UpdateClubDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}